Add file for json-server, edited App and package
This creates a new folder named db to hold the JSON file for the fake REST API of json-server. The App.vue handles this by creating a new async function named fetchTasks() which loads the data and populates it into the this.tasks array. Finally, we include new scripts in the package.json to include the backend launch script to start the json-server and a production script for the deployment.